Add email and name validation
